Changed the game options page for hearts. Users can add the number of players, there names, and how many points they are playing to. I also added logos for the app.
I also added a playerName variable to the Player class and updated the hearts game to show the player names instead of "Player 1", "Player 2", etc.

A user must enter the number of players, but other fields are optional. The default number of points is 50 and the default player names are "Player 1", "Player 2", etc.